Hack to get parity to sync faster
Implement a hack to get parity to sync faster in prod. This hack works
by first syncing locally, saving the synced blocks and embedding those
in a docker image. Then that image is pushed to gcr and the previously
synced blocks are used as a starting point. Ultimately we should aim to
use a mounted directory in prod a la https://app.clubhouse.io/mycro/story/63/create-persistent-volume-for-parity-in-gcp