Add TTL-LRU cache to github read requests
Add a ttl-lru cache to the github functions which perform reads. I saw
us hitting some api limits and decided to add this. This is better than
changing the poll interval in apollo because it will work when many
clients are viewing the website at once whereas a lower poll interval
won't.